Issue #1547: use adoc attributes for project
Add a static `project` attribute to the generated variables.adoc file
and refer to Minishift as a project with this attribute -- this allows
for easier project naming downstream.

Ensure variables.adoc is added to the troubleshooting section
directory.

Add `minikube-iso-version` attribute to variables.adoc.

Amend CLI profile reference in openshift-client-binary.adoc.