Adjust ocale to fix the string based network check
Previously we were checking for the (in other languages localized)
string 'active' in order to determine if the default network is active.
This failed whenever the language of the terminal had a translation which
was not the string 'active'.
This patch explicitly sets the language in order to fix this check.