Update instructions for using hyperkit
I tried using minishift with hyperkit and the existing documentation was unclear on which dependencies were provided by having Docker Mac installed. I reorganized that section a little to make it clear that there are two parts and that both of them can be satisfied from Homebrew but Docker Mac users only need docker-machine-driver-hyperkit.